requirements.txt
streamlit
pandas
Pillow
janome
matplotlib
japanize-matplotlib
fastapi
python-multipart